Re-organize bot code
