+++ src/db/mod.rs
@@ -1,0 +1,13 @@
+use sqlx::PgPool;
+use std::env;
+
+use crate::OddbotError;
+
+/// Create a new database pool
+pub async fn create_db_pool() -> Result<PgPool, OddbotError> {
+    let database_url = env::var("DATABASE_URL").expect("DATABASE_URL must be set");
+
+    PgPool::connect(&database_url)
+        .await
+        .map_err(OddbotError::Database)
+}
