+++ src/discord/mod.rs
@@ -1,0 +1,3 @@
+pub mod bot;
+pub mod events;
+pub mod handler;
